refactor(agents): split orchestrator module
